added middleware

+++ models/event.go
@@ -2,7 +2,6 @@ package models
 
 import (
 	"time"
-
 	"github.com/hkm007/gopherly/db"
 )
 
@@ -12,7 +11,7 @@ type Event struct {
 	Description string 	  `binding:"required"`
 	Location    string    `binding:"required"`
 	DateTime    time.Time `binding:"required"`
-	UserID      int
+	UserID      int64
 }
 
 func (e *Event) Save() error {
